COBRA-2944 - further isolation of dependencies
